refactor: extract quit dialog to src/ui/dialogs/quit_dialog.rs
- Move confirm_and_maybe_quit to ui/dialogs module
- Add OnDialogClose callback type for hotkey reinstall
- Add reinstall_hotkeys_callback wrapper in main.rs
- Remove duplicate function from main.rs (-311 lines)
- Total main.rs reduction: 689 lines (32% from original 2174)
